+++ src/main/java/jwp/controller/qna/QuestionFormController.java
@@ -1,6 +1,5 @@
-package jwp.controller;
+package jwp.controller.qna;
 
-import core.db.user.DBUserRepository;
 import core.mvc.Controller;
 import jwp.util.UserSessionUtils;
 
@@ -8,14 +7,13 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
-public class ListUserController implements Controller {
+public class QuestionFormController implements Controller {
 
     @Override
     public String execute(HttpServletRequest req, HttpServletResponse resp) throws Exception {
         HttpSession session = req.getSession();
         if (UserSessionUtils.isLogined(session)) {
-            req.setAttribute("users", DBUserRepository.getInstance().findAll());
-            return "/user/list.jsp";
+            return "/qna/form.jsp";
         }
         return "redirect:/user/loginForm";
     }
