+++ src/main/java/jwp/controller/user/LogoutController.java
@@ -1,4 +1,4 @@
-package jwp.controller;
+package jwp.controller.user;
 
 import core.mvc.Controller;
