+++ src/main/java/jwp/controller/user/UpdateUserController.java
@@ -1,4 +1,4 @@
-package jwp.controller;
+package jwp.controller.user;
 
 import core.db.user.DBUserRepository;
 import core.mvc.Controller;
@@ -7,16 +7,16 @@ import jwp.model.User;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-public class CreateUserController implements Controller {
+public class UpdateUserController implements Controller {
 
     @Override
     public String execute(HttpServletRequest req, HttpServletResponse resp) throws Exception {
-        User user = new User(req.getParameter("userId"),
+        User modifiedUser = new User(
+                req.getParameter("userId"),
                 req.getParameter("password"),
                 req.getParameter("name"),
                 req.getParameter("email"));
-        DBUserRepository.getInstance().addUser(user);
-        System.out.println("user 회원가입 완료");
+        DBUserRepository.getInstance().update(modifiedUser);
         return "redirect:/user/list";
     }
 }
