+++ src/main/java/jwp/controller/user/UpdateUserFormController.java
@@ -1,4 +1,4 @@
-package jwp.controller;
+package jwp.controller.user;
 
 import core.db.user.DBUserRepository;
 import core.mvc.Controller;
